requirements.txt
pandas
numpy
faker
pyyaml